10-8 Policy Authorization in Controllers

+++ app/Http/Controllers/JobController.php
@@ -3,15 +3,18 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use Illuminate\Foundation\Auth\Access\Authorizable;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -90,6 +93,9 @@ class JobController extends Controller
      */
     public function edit(Job $job): View
     {
+         // Check if user is authorized
+        $this->authorize('update', $job);
+        
         // return view('jobs.edit', compact('id'));
         return view('jobs.edit')->with('job', $job);
     }
@@ -99,6 +105,9 @@ class JobController extends Controller
      */
     public function update(Request $request, Job $job)
     {
+        // Check if user is authorized
+        $this->authorize('update', $job);
+
         $validatedData = $request->validate([
             'title' => ['required', 'string', 'max:255'],
             'description' => ['required', 'string'],
@@ -142,6 +151,9 @@ class JobController extends Controller
      */
     public function destroy(Job $job)
     {
+        // Check if user is authorized
+        $this->authorize('delete', $job);
+
         // If logo, then delete it
         if($job->logo) {
             Storage::delete('public/logos' . $job->company_logo);
